improve transpose

+++ tests/testthat/test-op-transpose.R
@@ -53,16 +53,16 @@ test_that("identity permutation", {
 test_that("scalar transpose", {
   skip_if_not_installed("pjrt")
   local_func()
-  x <- hlo_input("x", "f64", shape = integer())  # scalar (0 dimensions)
-  y <- hlo_transpose(x, permutation = integer())  # empty permutation for scalar
+  x <- hlo_input("x", "f64", shape = integer()) # scalar (0 dimensions)
+  y <- hlo_transpose(x, permutation = integer()) # empty permutation for scalar
   f <- hlo_return(y)
 
   program <- pjrt_program(repr(f))
   exec <- pjrt_compile(program)
 
-  input <- 42.0  # scalar value
-  expected <- input  # transpose of scalar should return the same value
+  input <- 42.0 # scalar value
+  expected <- input # transpose of scalar should return the same value
 
-  output <- pjrt_execute(exec, pjrt_buffer(input))
+  output <- pjrt_execute(exec, pjrt_buffer(input, dtype = "f64"))
   expect_equal(as_array(output), expected)
 })
